Specialized version of several find functions and `hash`.
To demonstrate the difference, `hash` is called for different sparse matrices. `hash` is calling `findprev` multiple times. 

Before - the time is growing with decreasing fill ratio.
```
julia> n = 10^5;
julia> A = sprand(rng, n, n, 10^8/n^2);
julia> @btime hash(A)
  129.376 ms (1 allocation: 16 bytes)
0x4c30d5f8ae24e7c4

julia> A = sprand(rng, n, n, 10^7/n^2);
julia> @btime hash(A)
  788.578 ms (1 allocation: 16 bytes)
0x7143db6e266eb8d8

julia> A = sprand(rng, n, n, 10^6/n^2);
julia> @btime hash(A)
  5.574 s (1 allocation: 16 bytes)
0x1ff7a9421112310d

julia> A = sprand(rng, n, n, 10^5/n^2);
julia> @btime hash(A)
  37.960 s (1 allocation: 16 bytes)
0x1083a7abc61aa2d2
```
After - the time is decreasing with decreasing fill ratio and generally lower.
```
julia> n = 10^5;
julia> A = sprand(rng, n, n, 10^8/n^2);
julia> @btime hash(A)
  29.570 ms (1 allocation: 16 bytes)
0x4c30d5f8ae24e7c4

julia> A = sprand(rng, n, n, 10^7/n^2);
julia> @btime hash(A)
  17.184 ms (1 allocation: 16 bytes)
0x7143db6e266eb8d8

julia> A = sprand(rng, n, n, 10^6/n^2);
julia> @btime hash(A)
  10.305 ms (1 allocation: 16 bytes)
0x1ff7a9421112310d

julia> A = sprand(rng, n, n, 10^5/n^2);
julia> @btime hash(A)
  8.319 ms (1 allocation: 16 bytes)
0x1083a7abc61aa2d2

julia> A = sprand(rng, n, n, 10^4/n^2);
julia> @btime hash(A)
  1.415 ms (1 allocation: 16 bytes)
0x02799c131c82e605
```