@KlausC 2019-03-15
AppVeyor fault seems unrelated.