@ViralBShah 2019-03-17
@klausC All your PRs seem to have a lot of commits that seem to be unrelated and clearly don't show up when I merge. 

I wonder if you are doing something differently with `git`, which produces these massive commit histories and messages. Perhaps someone knows what the issue is.