@yuyichao 2019-03-17
He just always use merge to sync with master and never rebased.