@KlausC 2019-03-17
@yuyichao That is right. I would like to insert a git rebase, if I'd know how to do that.
That is my current workstream: 
```
git checkout master
git fetch upstream
git checkout master
git merge upstream/master
git branch newbranch
git checkout newbranch
... edit ...
git push ...
```