@ViralBShah 2019-03-17
When you are ready to submit the PR, you may want to do `git rebase master` on your branch. Not sure if that is sufficient to fix this. You can try it here and force push to your branch.