@mauro3 2019-03-18
The workflow should be something like:
```
git checkout master
git pull # now your master is the same as upstream master.  Don't edit your master.
git checkout -b krc/newbranch
# work work

# if you need to get changes from upstream/master (only when there are conflicts):
git fetch upstream
git rebase upstream/master
```
Once done with a PR it is often nice to condense the PR into one or several important commits (each commit should pass the tests).  For this use `git rebase -i 123` where 123 is the sha of the commit just before you started work (if I recall correctly).

To get out of your current affliction, it's probably easiest if you copy your ` stdlib/SparseArrays/src/sparsematrix.jl` to somewhere, then do:
```
cp stdlib/SparseArrays/src/sparsematrix.jl /tmp/
git branch krc/findnext-backup # make a backup-branch in case something goes haywire
git fetch upstream/master
git reset --hard upstream/master # now your krc/findnext branch is identical to upstream/master
                                 # (be careful with reset --hard, it destroys stuff...)
cp /tmp/sparsematrix.jl stdlib/SparseArrays/src/sparsematrix.jl
git commit -am "..."
git push -f # fixes what's here on github
```