@KlausC 2019-03-18
@mauro3 , thank you very much! git looks clean now. I will change my work flow according to your recommendation.