@ViralBShah 2019-03-24
Can we have some tests for the sparse cases for findprev/next/etc?

Also, while it basically looks good to me, it would be good to get an extra pair of eyes on this PR.