@KlausC 2019-03-25
> Can we have some tests for the sparse cases for findprev/next/etc?

I thought, the existing tests might be sufficient:
hash: ( especially the critical -0.0 case is included) 
https://github.com/JuliaLang/julia/blob/10141d32ca15f76623236a5ddcd073ea650f146e/test/hashing.jl#L126-L147

findnext/findprev:
 https://github.com/JuliaLang/julia/blob/10141d32ca15f76623236a5ddcd073ea650f146e/stdlib/SparseArrays/test/sparse.jl#L2248-L2258