@KlausC 2019-03-25
> Also, while it basically looks good to me, it would be good to get an extra pair of eyes on this PR.

@andreasnoack, who would be interested?