@nalimilan 2019-03-25
AFAICT tests should include something like `findnext(iszero, y, k)` to cover the`if pred(zero(eltype(A)))` path. I also suspect they should test a matrix with stored zeros (both `0.0` and `-0.0`).