@KlausC 2019-03-26
Thanks for the review - I added the missing tests and made recommended changes.