@mbauman 2019-04-03
Yes, this looks quite good to me.  My only review comment would be in the naming of some of those `_` functions (I want a slightly better word for that index into `nzval` than `idx`), but they're underscore functions and I don't have any great ideas so I suppose I can't complain too much. :)

Thanks @KlausC!