@tkluck 2019-05-11
Hi there -- this pull request seems related to a noticeable slowdown I'm observing in my own use case. Could you help me see if we can fix that?

My use case is sparse matrices of polynomials, but the same issue exists for sparse matrices of e.g. `BigInt` or any other type that exists on the heap. The reason is that this PR introduces the following:

 -  `z = zero(<type>); ....; isequal(z, <something>)`.

Any reason not to use `iszero`?

 - Replacing [dispatch based on `f::typeof(!iszero)`](https://github.com/JuliaLang/julia/blob/7f2fa84e492af31cff31a6e46f676116a5c48dcc/stdlib/SparseArrays/src/abstractsparse.jl#L69) by an in-code check for `pred(zero(<type>))`. This latter thing is much more general, but in the case of `BigInt` (or my personal case of polynomials), it can't currently be elided by the compiler.

In my case, both of these things work together in such a way that the allocation of `zero(<type>)` is now by far the slowest part of my `findnext` calls. Eyeballing the flamegraph says it takes 80-90% of the time:

![flamegraph of findnext](https://user-images.githubusercontent.com/2110529/57570400-b3181500-7401-11e9-93d2-5513b0195219.png)

Disclose, I wrote the previous incarnations of this in #23317, so in that sense it's not a surprise that my use case works better with the code I optimized for it. Would you mind helping me to make both cases work?

Two more profiling screenshots to illustrate this:

![_idxnextnz](https://user-images.githubusercontent.com/2110529/57570446-4f421c00-7402-11e9-80b0-71c6d3601760.png)

(FWIW, this looks like it fulfills a very similar function to `_sparse_findnextnz`.)

![findnext](https://user-images.githubusercontent.com/2110529/57570439-30438a00-7402-11e9-86dc-ed238a531d46.png)