@tkluck 2019-05-11
It actually might make sense to look a bit deeper and see what's the difference between this PR and #23317 at all. It seems like they're both very similar and duplicate behaviour. My gut feeling is that it should have been possible to reach the stated objective of hash performance with only minor fixes to that other one.